+++ apps/client/src/components/errors/ErrorBlock.tsx
@@ -1,5 +1,5 @@
 import { PropsWithChildren } from "react";
-import { Button } from "../ui/button";
+import { Button } from "@/components/ui/button";
 import { isCriticalError } from "@/lib/errors";
 
 type TErrorBlockProps = {
