+++ apps/client/src/features/products/components/product-card/product-description.tsx
@@ -1,4 +1,4 @@
-import { UseProductCardContext } from "./index";
+import { UseProductCardContext } from "@/features/products/components/product-card/index";
 import { cn } from "@/lib/cn";
 
 export default function ProductDescription(props: { classes?: string }) {
